refactor: テストファイルをソースファイルと同じ場所に配置（コロケーション）
__tests__/ ディレクトリを廃止し、各テストファイルをテスト対象のソースファイルと
同じディレクトリに移動。インポートパスを ../→./ に修正。
